Rename integration tests to e2e tests
Preparing ground for more changes down the line. Renaming integration to e2e will create space for actual integration tests. The Cucumber tests will be simplified to adhere to design principles of the framework and they'll serve the purpose of e2e tests. The more detailed testing will be moved to integration tests not written in Cucumber.
