+++ test/e2e/features/story_registry.feature
